[#1]refactor: 코드 수정 - 필요없는 파일 삭제 - validation 추가 - 로직 분리 - 패키지 분리
