+++ src/main/java/com/spartatest/auth/application/AccountService.java
@@ -1,11 +1,9 @@
 package com.spartatest.auth.application;
 
-import com.spartatest.auth.domain.entity.Role;
 import com.spartatest.auth.domain.entity.User;
 import com.spartatest.auth.domain.repository.UserRepository;
-import com.spartatest.auth.dto.GrantAdminResponse;
-import com.spartatest.auth.dto.SignupRequest;
-import com.spartatest.auth.dto.SignupResponse;
+import com.spartatest.auth.dto.request.SignupRequest;
+import com.spartatest.auth.dto.response.SignupResponse;
 import com.spartatest.common.exception.CustomException;
 import com.spartatest.common.exception.custom.UserErrorCode;
 import lombok.RequiredArgsConstructor;
@@ -36,18 +34,5 @@ public class AccountService {
             throw new CustomException(UserErrorCode.USER_NOT_FOUND);
         }
         userRepository.deleteById(userId);
-    }
-
-    public GrantAdminResponse grantAdminRole(String targetUsername, User currentUser) {
-        if (!currentUser.getRole().equals(Role.ADMIN)) {
-            throw new CustomException(UserErrorCode.ACCESS_DENIED);
-        }
-
-        User user = userRepository.findByUsername(targetUsername)
-                .orElseThrow(() -> new CustomException(UserErrorCode.USER_NOT_FOUND));
-
-        user.grantAdminRole();
-        userRepository.save(user);
-        return GrantAdminResponse.from(user);
     }
 }
