--- src/main/java/com/spartatest/auth/dto/LoginRequest.java
@@ -1,6 +1,0 @@
-package com.spartatest.auth.dto;
-
-public record LoginRequest(
-        String username,
-        String password
-) {}
