--- src/main/java/com/spartatest/auth/dto/SignupRequest.java
@@ -1,7 +1,0 @@
-package com.spartatest.auth.dto;
-
-public record SignupRequest(
-        String username,
-        String password,
-        String nickname
-) {}
