--- src/main/java/com/spartatest/auth/dto/SignupResponse.java
@@ -1,14 +1,0 @@
-package com.spartatest.auth.dto;
-
-import com.spartatest.auth.domain.entity.Role;
-import com.spartatest.auth.domain.entity.User;
-
-public record SignupResponse(
-        String username,
-        String nickname,
-        Role roles
-) {
-    public static SignupResponse from(User user) {
-        return new SignupResponse(user.getUsername(), user.getNickname(), user.getRole());
-    }
-}
