+++ src/main/java/com/spartatest/auth/dto/response/GrantAdminResponse.java
@@ -1,0 +1,18 @@
+package com.spartatest.auth.dto.response;
+
+import com.spartatest.auth.domain.entity.Role;
+import com.spartatest.auth.domain.entity.User;
+
+public record GrantAdminResponse(
+        String username,
+        String nickname,
+        Role role
+) {
+    public static GrantAdminResponse from(User user) {
+        return new GrantAdminResponse(
+                user.getUsername(),
+                user.getNickname(),
+                user.getRole()
+        );
+    }
+}
