+++ src/main/java/com/spartatest/auth/dto/response/LoginResponse.java
@@ -1,6 +1,8 @@
-package com.spartatest.auth.dto;
+package com.spartatest.auth.dto.response;
 
-public record LoginResponse(String token) {
+public record LoginResponse(
+        String token
+) {
     public static LoginResponse from(String accessToken) {
         return new LoginResponse(accessToken);
     }
